STYLE: Apliquem correccions d'espais i salts de línia

+++ starviewer/src/inputoutput/converttodicomdir.cpp
@@ -39,7 +39,7 @@ void ConvertToDicomdir::setAnonymizeDICOMDIR(bool anonymizeDICOMDIR, QString pat
     m_patientNameAnonymized = patientNameAnonymized;
 }
 
-void ConvertToDicomdir::addStudy( const QString &studyUID )
+void ConvertToDicomdir::addStudy(const QString &studyUID)
 {
     /*Els estudis s'han d'agrupar per pacient, el que fem és afegir-los a llista d'estudis per convertir a
      dicomdir ja ordenats per pacient*/
@@ -51,7 +51,7 @@ void ConvertToDicomdir::addStudy( const QString &studyUID )
     DicomMask studyMask;
     studyMask.setStudyInstanceUID(studyUID);
     QList<Patient*> patientList = localDatabaseManager.queryPatientStudy(studyMask);
-    if(localDatabaseManager.getLastError() != LocalDatabaseManager::Ok)
+    if (localDatabaseManager.getLastError() != LocalDatabaseManager::Ok)
     {
         ERROR_LOG(QString("Error al afegir un study per generar un DICOMDIR; Error: %1; StudyUID: %2")
                   .arg(localDatabaseManager.getLastError())
@@ -68,26 +68,26 @@ void ConvertToDicomdir::addStudy( const QString &studyUID )
 
     delete patient;
 
-    while ( index < m_studiesToConvert.count() && !stop )  //busquem la posició on s'ha d'inserir l'estudi a llista d'estudis per convertir a dicomdir, ordenant per id de pacient
+    while (index < m_studiesToConvert.count() && !stop)  //busquem la posició on s'ha d'inserir l'estudi a llista d'estudis per convertir a dicomdir, ordenant per id de pacient
     {
-        if ( studyToConvert.patientId < m_studiesToConvert.at( index ).patientId ) //comparem amb els altres estudis de la llista, fins trobar el seu llloc corresponentm
+        if (studyToConvert.patientId < m_studiesToConvert.at(index).patientId) //comparem amb els altres estudis de la llista, fins trobar el seu llloc corresponentm
         {
             stop = true;
         }
         else index++;
     }
 
-    if ( stop ) //una vegada hem trobat la posició on ha d'anar l'inserim
+    if (stop) //una vegada hem trobat la posició on ha d'anar l'inserim
     {
-        m_studiesToConvert.insert( index , studyToConvert );
+        m_studiesToConvert.insert(index, studyToConvert);
     }
-    else m_studiesToConvert.push_back( studyToConvert );//en aquest cas val al final
+    else m_studiesToConvert.push_back(studyToConvert);//en aquest cas val al final
 }
 
 /*TODO:Si la creació del DICOMDIR Falla aquest mètode esborra el contingut del directori on s'havia de crear el DICOMDIR, al fer això
        la classe abans de crear el DICOMDIR hauria de comprovar que el directori està buit, perquè sinó podríem eliminar contingut del usuari.
        Ara aquesta comprovació es fa a la QCreateDicomdir i s'hauria de fer aquí*/
-Status ConvertToDicomdir::convert( const QString &dicomdirPath, CreateDicomdir::recordDeviceDicomDir selectedDevice, bool copyFolderContent)
+Status ConvertToDicomdir::convert(const QString &dicomdirPath, CreateDicomdir::recordDeviceDicomDir selectedDevice, bool copyFolderContent)
 {
     /* Primer copiem els estudis al directori desti, i posteriorment convertim el directori en un dicomdir*/
     Status state;
@@ -108,11 +108,11 @@ Status ConvertToDicomdir::convert( const QString &dicomdirPath, CreateDicomdir::
 
     LocalDatabaseManager localDatabaseManager;
 
-    foreach(StudyToConvert studyToConvert, m_studiesToConvert)
+    foreach (StudyToConvert studyToConvert, m_studiesToConvert)
     {
         studyMask.setStudyInstanceUID(studyToConvert.studyUID);
         Patient *patient = localDatabaseManager.retrieve(studyMask);
-        if(localDatabaseManager.getLastError() != LocalDatabaseManager::Ok)
+        if (localDatabaseManager.getLastError() != LocalDatabaseManager::Ok)
         {
             ERROR_LOG(QString("No s'han trobat les dades del estudi %1 a la base de dades ").arg(studyToConvert.studyUID));
             QString error = QString("Error al fer un retrieve study per generar un DICOMDIR; Error: %1; StudyUID: %2")
@@ -130,24 +130,24 @@ Status ConvertToDicomdir::convert( const QString &dicomdirPath, CreateDicomdir::
         // Nosaltres, en realitat només en volem un sol study.
         Study *study = patient->getStudies().first();
         studyList.append(study);
-        foreach(Series *series, study->getSeries())
+        foreach (Series *series, study->getSeries())
         {
             totalNumberOfItems += series->getNumberOfItems();
         }
     }
 
-    if ( !state.good() )
+    if (!state.good())
     {
-        DeleteDirectory().deleteDirectory( m_dicomDirPath , false );
+        DeleteDirectory().deleteDirectory(m_dicomDirPath, false);
         return state;
     }
 
     //sumem una imatge més per evitar que arribi el 100 % la progress bar, i així s'esperi a que es crei el dicomdir, que es fa quan s'invoca createDicomdir.Create()
-    m_progress = new QProgressDialog( tr( "Creating DICOMDIR..." ) , "" , 0 , totalNumberOfItems+1 );
-    m_progress->setMinimumDuration( 0 );
-    m_progress->setCancelButton( 0 );
+    m_progress = new QProgressDialog(tr("Creating DICOMDIR..."), "", 0, totalNumberOfItems + 1);
+    m_progress->setMinimumDuration(0);
+    m_progress->setCancelButton(0);
     m_progress->setModal(true);
-    
+
     if (m_anonymizeDICOMDIR)
     {
         INFO_LOG("El DICOMDIR es creara anonimitzat");
@@ -165,18 +165,18 @@ Status ConvertToDicomdir::convert( const QString &dicomdirPath, CreateDicomdir::
     //copiem les imatges dels estudis seleccionats al directori desti
     state = copyStudiesToDicomdirPath(studyList);
 
-    if ( !state.good() )
+    if (!state.good())
     {
         m_progress->close();
-        DeleteDirectory().deleteDirectory( m_dicomDirPath , false );
+        DeleteDirectory().deleteDirectory(m_dicomDirPath, false);
         return state;
     }
     //una vegada copiada les imatges les creem
-    state = createDicomdir( m_dicomDirPath , selectedDevice );
+    state = createDicomdir(m_dicomDirPath, selectedDevice);
 
-    if ( !state.good() && state.code() != 4001 )// l'error 4001 és que les imatges no compleixen l'estàndard al 100, però el dicomdir es pot utilitzar
+    if (!state.good() && state.code() != 4001)// l'error 4001 és que les imatges no compleixen l'estàndard al 100, però el dicomdir es pot utilitzar
     {
-        DeleteDirectory().deleteDirectory( m_dicomDirPath , false );
+        DeleteDirectory().deleteDirectory(m_dicomDirPath, false);
     }
     else
     {
@@ -188,7 +188,7 @@ Status ConvertToDicomdir::convert( const QString &dicomdirPath, CreateDicomdir::
             if (!copyFolderContentToDICOMDIR())
             {
                 m_progress->close();
-                DeleteDirectory().deleteDirectory( m_dicomDirPath, false);
+                DeleteDirectory().deleteDirectory(m_dicomDirPath, false);
                 state.setStatus("", false, 4002);
                 return state;
             }
@@ -214,7 +214,7 @@ Status ConvertToDicomdir::convert( const QString &dicomdirPath, CreateDicomdir::
     return state;
 }
 
-Status ConvertToDicomdir::createDicomdir( const QString &dicomdirPath, CreateDicomdir::recordDeviceDicomDir selectedDevice )
+Status ConvertToDicomdir::createDicomdir(const QString &dicomdirPath, CreateDicomdir::recordDeviceDicomDir selectedDevice)
 {
     CreateDicomdir createDicomdir;
     Status state, stateNotDicomConformance;
@@ -227,21 +227,21 @@ Status ConvertToDicomdir::createDicomdir( const QString &dicomdirPath, CreateDic
      ATENCIÓ si deshabilitem la comprovació de la transfer syntax podem tenir DICOMDIR que no siguin DICOM conformance*/
     createDicomdir.setCheckTransferSyntax(getConvertDicomdirImagesToLittleEndian());
 
-    createDicomdir.setDevice( selectedDevice );
-    state = createDicomdir.create( dicomdirPath );//invoquem el mètode per convertir el directori destí Dicomdir on ja s'han copiat les imatges en un dicomdir
-    if ( !state.good() )//ha fallat crear el dicomdir, ara intentem crear-lo en mode no estricte
+    createDicomdir.setDevice(selectedDevice);
+    state = createDicomdir.create(dicomdirPath);//invoquem el mètode per convertir el directori destí Dicomdir on ja s'han copiat les imatges en un dicomdir
+    if (!state.good())//ha fallat crear el dicomdir, ara intentem crear-lo en mode no estricte
     {
         WARN_LOG("Algunes de les imatges no compleixen l'estandard DICOM al 100% es provara de crear el DICOMDIR sense el mode estricte");
-        createDicomdir.setStrictMode( false );
-        state = createDicomdir.create( dicomdirPath );
-        if ( state.good() )
+        createDicomdir.setStrictMode(false);
+        state = createDicomdir.create(dicomdirPath);
+        if (state.good())
         {
-            return stateNotDicomConformance.setStatus("Alguna de les imatges no complia l'estàndard DICOM" , false , 4001 );
+            return stateNotDicomConformance.setStatus("Alguna de les imatges no complia l'estàndard DICOM", false, 4001);
         }
     }
 
     //Hem assignat com a valor de progressbar Numero Imatges +1, el +1 és el pas de convertir els fitxers a dicomdir
-    m_progress->setValue(m_progress->value() +1);
+    m_progress->setValue(m_progress->value() + 1);
 
     return state;
 }
@@ -284,7 +284,7 @@ Status ConvertToDicomdir::copyStudiesToDicomdirPath(QList<Study*> studyList)
     m_patient = 0;
 
     //agrupem estudis1 per pacient, com que tenim la llista ordenada per patientId
-    while ( !m_studiesToConvert.isEmpty() )
+    while (!m_studiesToConvert.isEmpty())
     {
         studyToConvert = m_studiesToConvert.takeFirst();
 
@@ -293,27 +293,27 @@ Status ConvertToDicomdir::copyStudiesToDicomdirPath(QList<Study*> studyList)
         // de que hi tenim el mateix ordre.
         study = studyList.takeFirst();
 
-        if(study->getInstanceUID() != studyToConvert.studyUID)
+        if (study->getInstanceUID() != studyToConvert.studyUID)
         {
             state.setStatus("La xapussa del copyStudiesToDicomdirPath no funciona, hi ha ordre diferent entre la llista studyList i m_studiesToConvert", false, -1);
             break;
         }
 
         //si el pacient es diferent creem un nou directori PAtient
-        if ( m_OldPatientId != studyToConvert.patientId )
+        if (m_OldPatientId != studyToConvert.patientId)
         {
-            patientNameDir = QString( "/PAT%1" ).arg( m_patient , 5 , 10 , fillChar );
+            patientNameDir = QString("/PAT%1").arg(m_patient, 5, 10, fillChar);
             m_dicomdirPatientPath = m_dicomDirPath + "/DICOM/" + patientNameDir;
-            patientDir.mkpath( m_dicomdirPatientPath );
+            patientDir.mkpath(m_dicomdirPatientPath);
             m_patient++;
             m_study = 0;
-            m_patientDirectories.push_back( m_dicomdirPatientPath );//creem una llista amb els directoris creats, per si es produeix algun error esborrar-los
+            m_patientDirectories.push_back(m_dicomdirPatientPath);//creem una llista amb els directoris creats, per si es produeix algun error esborrar-los
         }
 
         state = copyStudyToDicomdirPath(study);
 
         delete study;
-        if ( !state.good() ) break;
+        if (!state.good()) break;
         m_OldPatientId = studyToConvert.patientId;
     }
 
@@ -335,9 +335,9 @@ Status ConvertToDicomdir::copyStudyToDicomdirPath(Study *study)
     m_dicomDirStudyPath = m_dicomdirPatientPath + studyName;
     studyDir.mkdir(m_dicomDirStudyPath);
 
-    foreach(Series *series, study->getSeries()) //per cada sèrie de l'estudi, creem el directori de la sèrie
+    foreach (Series *series, study->getSeries()) //per cada sèrie de l'estudi, creem el directori de la sèrie
     {
-        if ( series->getNumberOfItems() > 0 )
+        if (series->getNumberOfItems() > 0)
         {
             state = copySeriesToDicomdirPath(series);
 
@@ -353,26 +353,26 @@ Status ConvertToDicomdir::copySeriesToDicomdirPath(Series *series)
     QDir seriesDir;
     QChar fillChar = '0';
     //creem el nom del directori de la sèrie, el format és SERXXXXX, on XXXXX és el numero de sèrie dins l'estudi
-    QString seriesName = QString( "/SER%1" ).arg( m_series , 5 , 10 , fillChar );
+    QString seriesName = QString("/SER%1").arg(m_series, 5, 10, fillChar);
     Status state;
 
     m_series++;
     m_image = 0;
     //Creem el directori on es guardarà la sèrie en format DicomDir
     m_dicomDirSeriesPath = m_dicomDirStudyPath + seriesName;
-    seriesDir.mkdir( m_dicomDirSeriesPath );
+    seriesDir.mkdir(m_dicomDirSeriesPath);
 
     // HACK per evitar els casos en que siguin imatges procedents d'un multiframe
     // que copiem més d'una vegada un arxiu
     QString lastPath;
-    foreach(Image *imageToCopy, series->getImages())
+    foreach (Image *imageToCopy, series->getImages())
     {
-        if( lastPath != imageToCopy->getPath() )
+        if (lastPath != imageToCopy->getPath())
         {
             lastPath = imageToCopy->getPath();
-            state = copyImageToDicomdirPath( imageToCopy );
-            
-            if ( !state.good() ) 
+            state = copyImageToDicomdirPath(imageToCopy);
+
+            if (!state.good())
                 break;
         }
     }
@@ -383,14 +383,14 @@ Status ConvertToDicomdir::copySeriesToDicomdirPath(Series *series)
 Status ConvertToDicomdir::copyImageToDicomdirPath(Image *image)
 {
     //creem el nom del fitxer de l'imatge, el format és IMGXXXXX, on XXXXX és el numero d'imatge dins la sèrie
-    QString imageOutputPath = m_dicomDirSeriesPath + QString( "/IMG%1" ).arg( m_image , 5 , 10 , QChar('0'));
+    QString imageOutputPath = m_dicomDirSeriesPath + QString("/IMG%1").arg(m_image, 5, 10, QChar('0'));
     Status state;
     m_image++;
 
     if (getConvertDicomdirImagesToLittleEndian())
     {
         //convertim la imatge a littleEndian, demanat per la normativa DICOM i la guardem al directori desti
-        state = ConvertDicomToLittleEndian().convert(image->getPath(), imageOutputPath );
+        state = ConvertDicomToLittleEndian().convert(image->getPath(), imageOutputPath);
 
         if (m_anonymizeDICOMDIR && state.good())
         {
@@ -403,8 +403,8 @@ Status ConvertToDicomdir::copyImageToDicomdirPath(Image *image)
     }
     else
     {
-        //Si hem d'anonimitzar el fitxer el que fem és que en comptes de copiar-lo i llavors anonimitzar-lo, és  indicar-li al mètode d'anonimitzar 
-        //que guardi el fitxer en el lloc on s'hauria hagut de copiar per crear el DICOMDIR, d'aquesta manera la creació de DICOMDIR per imatges 
+        //Si hem d'anonimitzar el fitxer el que fem és que en comptes de copiar-lo i llavors anonimitzar-lo, és  indicar-li al mètode d'anonimitzar
+        //que guardi el fitxer en el lloc on s'hauria hagut de copiar per crear el DICOMDIR, d'aquesta manera la creació de DICOMDIR per imatges
         //que no s'han de convertir a LittleEndian és més ràpid.
         if (m_anonymizeDICOMDIR)
         {
@@ -418,7 +418,7 @@ Status ConvertToDicomdir::copyImageToDicomdirPath(Image *image)
         {
             if (QFile::copy(image->getPath(), imageOutputPath))
             {
-                state.setStatus("",true,0);
+                state.setStatus("", true, 0);
             }
             else
             {
@@ -428,7 +428,7 @@ Status ConvertToDicomdir::copyImageToDicomdirPath(Image *image)
         }
     }
 
-    m_progress->setValue( m_progress->value() + 1 ); // la barra de progrés avança
+    m_progress->setValue(m_progress->value() + 1); // la barra de progrés avança
     m_progress->repaint();
 
     return state;
@@ -438,9 +438,9 @@ void ConvertToDicomdir::createReadmeTxt()
 {
     Settings settings;
     QString readmeFilePath = m_dicomDirPath + "/README.TXT";
-    QFile file( readmeFilePath );
+    QFile file(readmeFilePath);
 
-    if (file.exists()) 
+    if (file.exists())
     {
         INFO_LOG("Ja s'ha trobat un fitxer readme.txt al DICOMDIR provinent del contingut de la carpeta a copiar");
         /*Si el fitxer ja existeix vol dir que l'hem copiat del contingut la carpeta que s'ha copiar al crear un DICOMDIR
@@ -448,17 +448,17 @@ void ConvertToDicomdir::createReadmeTxt()
         return;
     }
 
-    if ( !file.open( QIODevice::WriteOnly | QIODevice::Text ) ) return;
+    if (!file.open(QIODevice::WriteOnly | QIODevice::Text)) return;
 
-    QTextStream out( &file );
+    QTextStream out(&file);
     out << "The DICOMDIR has been generated with " << ApplicationNameString << " Version " << StarviewerVersionString << "\n";
     out << "E-mail contact : " << OrganizationEmailString << "\n";
     out << ApplicationNameString << " is not responsible for DICOMDIR content." << "\n\n";
 
     if (!settings.getValue(InputOutputSettings::InstitutionName).toString().isEmpty())
     {
-        out << "The DICOMDIR has been created by: " << settings.getValue(InputOutputSettings::InstitutionName).toString() << "\n"; 
-        
+        out << "The DICOMDIR has been created by: " << settings.getValue(InputOutputSettings::InstitutionName).toString() << "\n";
+
         if (!settings.getValue(InputOutputSettings::InstitutionAddress).toString().isEmpty())
         {
             out << settings.getValue(InputOutputSettings::InstitutionAddress).toString() << "\n";
